Added the AwesomePoints property. This is updated when the game state is updated.
Removed the SelectedCard behavior from the GameActivity. This won't work when the horizontal list view keeps regenerating the views for the cards.

Changed the logic for the NewData method. Instead of replacing the list, I only remove the old cards and add the new ones. I use the Id to determine which cards have changed.