Changed the CardsAgainstHumanity logo to CardsAgainstHumility in the BlackCard and WhiteCard layouts
Added logic for the CardCzar player gameplay.

Added a button to "Ready Up" at the end of the round.

Changed the socket logic to only have one socket, and to switch between the lobby and the game with one socket.

Moved all the gamestate update logic into one method that's called at the end of the async calls and when receiving updateGame events from the socket.

The Game socket is not connecting for some reason. I need to figure this out. It was connecting this morning.