Fixed some issues with inconsistent rendering of cards on Marshmallow compared to earlier versions of android.
Converted the cardCzar notification text at the top of the game view to be a general status field that notifies the player of what is going on while waiting on other players.

Fixed the game socket issue by changing the address on the server from '/' to '/game' (this fixed it for the web app as well)

Separated the port from the server address, though that wasn't really 100% necessary

Fixed an issue on the server that would cause the departgame task to never complete. This caused an appcrash even after the user had 'closed' the app.