Changed UserInfo back to using the Profile DisplayName instead of the weird number thing it was using.
Removed the port field from the Settings screen and the corresponding property from the CardsAgainstHumility static class. Changed the logic to account for this.

Changed the default server address from an IP address to the url of the new Heroku app where I'm hosting it. Performance should be ... different ... now.