Changed the placement of white cards to be just below the black card on screen when showing selected cards.
Changed the game to hide the player hand or played cards when showing a selected card with the current black card.

Added a check for isFinishing before departing the game in the OnDestroy. Apparently OnDestroy is called when the screen is rotated, and the user doesn't need to depart from the game unless they're changing to a different activity.

Removed the async tag from the DepartGame method