Removed the "Get User Name" logic in favor of just defaulting to "Smelly Idiot"
Added faint outlines tothe BlackCard and WhiteCard xml vector graphics. These allow the cards to stand out when rendered atop other cards of the same type.

Greatly simplified the code for determining if the player hand and selected card views should be visible at any given time. Split the SelectedCardView logic into a separate method.

Fixed the bug that caused the cards not to be immediately visible on the screen when the card czar is finally given access to the played cards for the round.

Added a close button to the RoundWinnerAnnouncement so that the user knows its a modal that needs to be closed before continuing.