Moved the project into its own folder. I was going to move most of the gameplay code into a separate portable class library, but the socket.io stuff would have had to be made into an interface and implemented in platform specific logic, and that would have been gross.
Added a SettingsLoader interface and implemented it on the android side so that CardsAgainstHumility wouldn't need any android stuff there.

Added an arrow button midway down the left side of the screen so Eric can open the player list without a lot of effort.
Moved the PrepareWhiteCard and PrepareBlackCard methods into UIAssets and put that in its own namespace. Replaced the logic in the white card array adapter with a call to preparewhitecard instead.